+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- if a pipeline fails, the container is kept regardless of `--keep` or not
+- `--verbatim` option to only output from pipeline, not pipelines
 ### Changed
 - Various code style improvements
-### Added
-- `--verbatim` option to only output from pipeline, not pipelines
 
 ## [0.0.2] - 2018-01-11
 ### Added